+++ tetris/constant.js
@@ -1,0 +1,7 @@
+export const CONSTENT = {
+  STROKE: "#222222",
+  FILL: "#2962ff",
+  LINE_WIDTH: 1,
+  SIDE_LENGTH: 10,
+  STEP: 5,
+};
