+++ tetris/index.js
@@ -1,23 +1,9 @@
-const UTILS = {
-  log: console.log.bind(console),
-  e(s) {
-    return document.querySelectorAll(s);
-  },
-  $(s) {
-    return this.e(s)[0];
-  },
-  $s(s) {
-    return this.e(s);
-  },
-};
-
-const CONSTENT = {
-  STROKE: "#222222",
-  FILL: "#2962ff",
-  LINE_WIDTH: 1,
-  SIDE_LENGTH: 10,
-  STEP: 5,
-};
+import IShape from "./shapes/IShape";
+import OShape from "./shapes/OShape";
+import SShape from "./shapes/SShape";
+import ZShape from "./shapes/ZShape";
+import { CONSTENT } from "./constant";
+import { UTILS } from "./utils";
 
 const KEY_CODES = {
   SPACE: 32,
@@ -38,270 +24,6 @@ context.scale(2, 2);
 context.strokeStyle = CONSTENT.STROKE;
 context.fillStyle = CONSTENT.FILL;
 context.lineWidth = CONSTENT.LINE_WIDTH;
-
-class Shape {
-  constructor(options) {
-    const { x, y, width, height } = options;
-
-    this.x = x;
-    this.y = y;
-    this.width = width;
-    this.height = height;
-  }
-
-  static getHeight(shape) {}
-
-  moveLeft() {
-    this.x -= CONSTENT.SIDE_LENGTH;
-  }
-
-  moveRight() {
-    this.x += CONSTENT.SIDE_LENGTH;
-  }
-
-  check() {
-    return false;
-  }
-
-  update() {
-    this.y += CONSTENT.SIDE_LENGTH;
-  }
-}
-
-class IShape extends Shape {
-  constructor(options) {
-    super({
-      ...options,
-      width: CONSTENT.SIDE_LENGTH * 4,
-      height: CONSTENT.SIDE_LENGTH * 1,
-    });
-
-    this.isHorizontal = true;
-  }
-
-  changeShape() {
-    this.isHorizontal = !this.isHorizontal;
-  }
-
-  getHeight() {
-    if (this.isHorizontal) {
-      return this.height;
-    } else {
-      return this.width;
-    }
-  }
-
-  render(ctx) {
-    let { x, y, width, height, isHorizontal } = this;
-
-    if (!this.isHorizontal && y > 200) {
-      y = 200;
-    }
-
-    let i = 0;
-    let c = CONSTENT.SIDE_LENGTH;
-    for (; i < width; i += c) {
-      ctx.beginPath();
-
-      if (isHorizontal) {
-        ctx.rect(x + i, y, c, c);
-      } else {
-        ctx.rect(x, y + i, c, c);
-      }
-
-      ctx.stroke();
-      ctx.fill();
-    }
-  }
-}
-
-class SShape extends Shape {
-  constructor(options) {
-    super({
-      ...options,
-      width: CONSTENT.SIDE_LENGTH * 3,
-      height: CONSTENT.SIDE_LENGTH * 2,
-    });
-
-    this.isTruned = false;
-  }
-
-  check(i, j) {
-    let c = CONSTENT.SIDE_LENGTH;
-    const { isTruned } = this;
-
-    if (isTruned) {
-      return (j === 0 && i === 0) || (j === c && i === 2 * c);
-    } else {
-      return (j === 0 && i === 0) || (j === c && i === 2 * c);
-    }
-  }
-
-  getHeight() {
-    if (this.isTruned) {
-      return this.height;
-    } else {
-      return this.height + CONSTENT.SIDE_LENGTH;
-    }
-    return;
-  }
-
-  changeShape() {
-    this.isTruned = !this.isTruned;
-  }
-
-  render(ctx) {
-    const { x, y, width, height, isTruned } = this;
-
-    let row = 0,
-      column = 0;
-    let c = CONSTENT.SIDE_LENGTH;
-
-    if (isTruned) {
-      for (; column < height; column += c) {
-        for (row = 0; row < width; row += c) {
-          if (this.check(row, column)) {
-            continue;
-          }
-
-          ctx.beginPath();
-
-          ctx.rect(x + row, y + column, c, c);
-
-          ctx.stroke();
-          ctx.fill();
-        }
-      }
-    } else {
-      for (; column < height + c; column += c) {
-        for (row = 0; row < width - c; row += c) {
-          if (row === 0 && column === 0) {
-            continue;
-          }
-          if (column === c * 2 && row === c) {
-            continue;
-          }
-          ctx.beginPath();
-
-          ctx.rect(x + row, y + column, c, c);
-
-          ctx.stroke();
-          ctx.fill();
-        }
-      }
-    }
-  }
-}
-
-class ZShape extends Shape {
-  constructor(options) {
-    super({
-      ...options,
-      width: CONSTENT.SIDE_LENGTH * 3,
-      height: CONSTENT.SIDE_LENGTH * 2,
-    });
-
-    this.isTruned = false;
-  }
-
-  check(i, j) {
-    let c = CONSTENT.SIDE_LENGTH;
-    const { isTruned } = this;
-
-    if (isTruned) {
-      return (j === 0 && i === 2 * c) || (j === c && i === 0);
-    } else {
-      return (j === 0 && i === 0) || (j === c && i === 2 * c);
-    }
-  }
-
-  getHeight() {
-    if (this.isTruned) {
-      return this.height;
-    } else {
-      return this.height + CONSTENT.SIDE_LENGTH;
-    }
-    return;
-  }
-
-  changeShape() {
-    this.isTruned = !this.isTruned;
-  }
-
-  render(ctx) {
-    const { x, y, width, height, isTruned } = this;
-
-    let row = 0,
-      column = 0;
-    let c = CONSTENT.SIDE_LENGTH;
-
-    if (isTruned) {
-      for (; column < height; column += c) {
-        for (row = 0; row < width; row += c) {
-          if (this.check(row, column)) {
-            continue;
-          }
-
-          ctx.beginPath();
-
-          ctx.rect(x + row, y + column, c, c);
-
-          ctx.stroke();
-          ctx.fill();
-        }
-      }
-    } else {
-      for (; column < height + c; column += c) {
-        for (row = 0; row < width - c; row += c) {
-          if (row === 0 && column === 0) {
-            continue;
-          }
-          if (column === c * 2 && row === c) {
-            continue;
-          }
-          ctx.beginPath();
-
-          ctx.rect(x + row, y + column, c, c);
-
-          ctx.stroke();
-          ctx.fill();
-        }
-      }
-    }
-  }
-}
-
-class OShape extends Shape {
-  constructor(options) {
-    super({
-      ...options,
-      width: CONSTENT.SIDE_LENGTH * 2,
-      height: CONSTENT.SIDE_LENGTH * 2,
-    });
-  }
-
-  getHeight() {
-    return this.height;
-  }
-
-  render(ctx) {
-    const { x, y, width, height } = this;
-
-    let row = 0,
-      column = 0;
-    let c = CONSTENT.SIDE_LENGTH;
-    for (; column < height; column += c) {
-      for (row = 0; row < width; row += c) {
-        ctx.beginPath();
-
-        ctx.rect(x + row, y + column, c, c);
-
-        ctx.stroke();
-        ctx.fill();
-      }
-    }
-  }
-}
 
 // shapes i l j o z s t
 
@@ -424,7 +146,7 @@ export function main() {
   }
 }
 
-UTILS.log(canvas);
+// UTILS.log(canvas);
 
 function reset() {
   paused = false;
