+++ tetris/utils.js
@@ -1,0 +1,12 @@
+export const UTILS = {
+  log: console.log.bind(console),
+  e(s) {
+    return document.querySelectorAll(s);
+  },
+  $(s) {
+    return this.e(s)[0];
+  },
+  $s(s) {
+    return this.e(s);
+  },
+};
